Feat/read-only and to-skip storage services

+++ src/Storage/Service/AbstractUrlStorage.php
@@ -4,15 +4,32 @@ namespace EMS\CommonBundle\Storage\Service;
 
 use GuzzleHttp\Psr7\Stream;
 use Psr\Http\Message\StreamInterface;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 abstract class AbstractUrlStorage implements StorageInterface
 {
+    /** @var bool */
+    private $readOnly;
+    /** @var bool */
+    private $skip;
+    /** @var LoggerInterface */
+    private $logger;
+
+    public function __construct(LoggerInterface $logger, bool $readOnly, bool $skip)
+    {
+        $this->logger = $logger;
+        $this->readOnly = $readOnly;
+        $this->skip = $skip || $readOnly;
+    }
 
     abstract protected function getBaseUrl(): string;
 
     protected function initDirectory(string $filename): void
     {
+        if ($this->isReadOnly()) {
+            return;
+        }
         if (!\file_exists(\dirname($filename))) {
             \mkdir(\dirname($filename), 0777, true);
         }
@@ -43,6 +60,9 @@ abstract class AbstractUrlStorage implements StorageInterface
 
     public function create(string $hash, string $filename): bool
     {
+        if ($this->isReadOnly()) {
+            return false;
+        }
         $path = $this->getPath($hash);
         $this->initDirectory($path);
         return copy($filename, $path);
@@ -92,6 +112,9 @@ abstract class AbstractUrlStorage implements StorageInterface
 
     public function remove(string $hash): bool
     {
+        if ($this->isReadOnly()) {
+            return false;
+        }
         $file = $this->getPath($hash);
         if (file_exists($file)) {
             unlink($file);
@@ -102,6 +125,9 @@ abstract class AbstractUrlStorage implements StorageInterface
 
     public function initUpload(string $hash, int $size, string $name, string $type): bool
     {
+        if ($this->isReadOnly()) {
+            return false;
+        }
         $path = $this->getUploadPath($hash);
         $this->initDirectory($path);
         return file_put_contents($path, "") !== false;
@@ -109,6 +135,9 @@ abstract class AbstractUrlStorage implements StorageInterface
 
     public function addChunk(string $hash, string $chunk): bool
     {
+        if ($this->isReadOnly()) {
+            return false;
+        }
         $path = $this->getUploadPath($hash);
         if (!file_exists($path)) {
             throw new NotFoundHttpException('temporary file not found');
@@ -132,19 +161,32 @@ abstract class AbstractUrlStorage implements StorageInterface
 
     public function finalizeUpload(string $hash): bool
     {
+        if ($this->isReadOnly()) {
+            return false;
+        }
         $source = $this->getUploadPath($hash);
         $destination  = $this->getPath($hash);
         $this->initDirectory($destination);
         try {
             return \rename($source, $destination);
         } catch (\Throwable $e) {
-            //TODO: add log info or notice
+            $this->logger->info(sprintf('Rename %s to %s failed: %s', $source, $destination, $e->getMessage()));
         }
         try {
             return \copy($source, $destination);
         } catch (\Throwable $e) {
-            //TODO: add log info or notice
+            $this->logger->info(sprintf('Copy %s to %s failed: %s', $source, $destination, $e->getMessage()));
         }
         return false;
+    }
+
+    public function isReadOnly(): bool
+    {
+        return $this->readOnly;
+    }
+
+    public function shouldSkip(): bool
+    {
+        return $this->skip;
     }
 }
